refactor(desktop): extract chat transcript logic to messages/
Verbatim moves out of chat-messages.tsx (2,415 -> ~1,300 lines), with no behavior changes.

- Extract shared constants, grouping and reasoning helpers, tool summaries, and tool icons into messages/.
- Add unit tests for the extracted pure logic.
- Update test:chat-ui to include tests under messages/.

chat-messages.test.tsx remains unchanged and continues to pass.
